requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
bitarray==2.8.1
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
colorama==0.4.6
crashtest==0.4.1
cytoolz==0.12.2
distlib==0.3.7
Django==4.2.2
django-autoslug==1.9.9
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dulwich==0.21.5
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.2
fastapi==0.101.0
filelock==3.12.2
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
hexbytes==0.3.1
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
lru-dict==1.2.0
magic-filter==1.0.11
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
parsimonious==0.9.0
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.24.0
ptyprocess==0.7.0
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
pygame==2.4.0
PyJWT==2.8.0
pyocr==0.8.3
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-multipart==0.0.6
python-nmap==0.7.1
python-telegram-bot==20.4
python-whois==0.8.0
pytz==2023.3
pyunormalize==15.0.0
pywin32==306
pywin32-ctypes==0.2.2
rapidfuzz==2.15.1
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rlp==3.0.0
rpds-py==0.9.2
shell.py==0.5.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
trove-classifiers==2023.7.6
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.24.2
web3==6.8.0
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
